Refactored Git utilities to `Kit` module.
- In order to improve gem organization and group similar objects
  together, the lower-level Git utility/wrapper classes have been moved
  to the `Kit` module.
