+++ lib/git/cop.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 require "git/cop/identity"
-require "git/cop/branch"
-require "git/cop/commit"
+require "git/cop/kit/branch"
+require "git/cop/kit/commit"
 require "git/cop/styles/abstract"
 require "git/cop/styles/commit_author_email"
 require "git/cop/styles/commit_author_name_capitalization"
