+++ spec/lib/git/cop/kit/commit_spec.rb
@@ -2,7 +2,7 @@
 
 require "spec_helper"
 
-RSpec.describe Git::Cop::Commit, :git_repo do
+RSpec.describe Git::Cop::Kit::Commit, :git_repo do
   let(:sha) { Dir.chdir(git_repo_dir) { `git log --pretty=format:%H -1` } }
   subject { Dir.chdir(git_repo_dir) { described_class.new sha: sha } }
