+++ spec/lib/git/cop/styles/commit_body_bullet_spec.rb
@@ -4,7 +4,7 @@ require "spec_helper"
 
 RSpec.describe Git::Cop::Styles::CommitBodyBullet do
   let(:body_lines) { ["- Test message."] }
-  let(:commit) { object_double Git::Cop::Commit.new(sha: "1"), body_lines: body_lines }
+  let(:commit) { object_double Git::Cop::Kit::Commit.new(sha: "1"), body_lines: body_lines }
   subject { described_class.new commit: commit }
 
   describe ".id" do
