+++ spec/lib/git/cop/styles/commit_body_leading_space_spec.rb
@@ -4,7 +4,7 @@ require "spec_helper"
 
 RSpec.describe Git::Cop::Styles::CommitBodyLeadingSpace do
   let(:raw_body) { "Added documentation.\n\n- Necessary for testing purposes.\n" }
-  let(:commit) { object_double Git::Cop::Commit.new(sha: "1"), raw_body: raw_body }
+  let(:commit) { object_double Git::Cop::Kit::Commit.new(sha: "1"), raw_body: raw_body }
   let(:settings) { {enabled: true} }
   subject { described_class.new commit: commit, settings: settings }
