+++ spec/lib/git/cop/styles/commit_body_present_spec.rb
@@ -4,7 +4,7 @@ require "spec_helper"
 
 RSpec.describe Git::Cop::Styles::CommitBodyPresent do
   let(:body_lines) { ["Curabitur eleifend wisi iaculis ipsum."] }
-  let(:commit) { object_double Git::Cop::Commit.new(sha: "1"), body_lines: body_lines }
+  let(:commit) { object_double Git::Cop::Kit::Commit.new(sha: "1"), body_lines: body_lines }
   let(:minimum) { 1 }
   let(:settings) { {enabled: true, minimum: minimum} }
   subject { described_class.new commit: commit, settings: settings }
