+++ spec/lib/git/cop/styles/commit_subject_length_spec.rb
@@ -4,7 +4,7 @@ require "spec_helper"
 
 RSpec.describe Git::Cop::Styles::CommitSubjectLength do
   let(:content) { "Added test subject." }
-  let(:commit) { object_double Git::Cop::Commit.new(sha: "1"), subject: content }
+  let(:commit) { object_double Git::Cop::Kit::Commit.new(sha: "1"), subject: content }
   let(:length) { 25 }
   let(:settings) { {enabled: true, length: length} }
   subject { described_class.new commit: commit, settings: settings }
